create new invoices

+++ dokuku.sales.invoices/Invoices.cs
@@ -24,13 +24,10 @@ namespace dokuku.sales.invoices
         public string TermCondition { get; set; }
         public decimal SubTotal { get; set; }
         public decimal Total { get; set; }
-
         public InvoiceItems[] Items { get; set; }
-
         public Guid _id { get; set; }
         public string _rev { get; set; }
         public string OwnerId { get; set; }
-
     }
 
     public class InvoiceItems
@@ -39,8 +36,8 @@ namespace dokuku.sales.invoices
         public string Description { get; set; }
         public int Qty { get; set; }
         public decimal Rate { get; set; }
-        public int Discount { get; set; }
-        public int Tax { get; set; }
+        public decimal Discount { get; set; }
+        public decimal Tax { get; set; }
         public decimal Amount { get; set; }     
         
     }
